product-service2 added

+++ orderService/src/main/java/miu/edu/sa/orderservice/model/Order.java
@@ -18,7 +18,7 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long orderId;
-    private int productId;
-    private int quantity;
+    private Long productId;
+    private Integer quantity;
     private String paymentMethod;
 }
